backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
drf-spectacular>=0.27,<0.28
django-filter>=24.3,<25.0
django-cors-headers>=4.4,<5.0
mysqlclient>=2.2,<2.3
redis>=5.0,<6.0
django-redis>=5.4,<5.5
